Extract functions from SwitchLevel capability
Related to #44
