+++ src/hubitat-capabilities/index.ts
@@ -5,6 +5,7 @@ export * from './functions/get-battery-status.function';
 export * from './functions/get-contact-sensor-status.function';
 export * from './functions/get-cooling-setpoint.function';
 export * from './functions/get-heating-setpoint.function';
+export * from './functions/get-level.function';
 export * from './functions/get-location-mode.function';
 export * from './functions/get-number-of-buttons.function';
 export * from './functions/get-supported-thermostat-fan-modes.function';
@@ -24,6 +25,7 @@ export * from './functions/is-present.function';
 export * from './functions/set-alarm-setting.function';
 export * from './functions/set-cooling-setpoint.function';
 export * from './functions/set-heating-setpoint.function';
+export * from './functions/set-level.funciton';
 export * from './functions/set-supported-thermostat-fan-modes.function';
 export * from './functions/set-supported-thermostat-modes.function';
 export * from './functions/set-switch.function';
